+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- `consent/prompt` accepts `client` and `scope` options via res.locals.
+- `application` and `scope` locals are set prior to rendering 'consent' view.
+- 'consent' prompt accepts `client` and `scope` options via res.locals.
 
 ### Changed